[Utils refactor] Inline array utils
