--- packages/core/utils/src/arrayUtils.ts
@@ -1,22 +1,0 @@
-function arrayRemove<T>(array: T[], item: T) {
-  const updatedArray = [...array];
-  const index = updatedArray.indexOf(item);
-  if (index !== -1) {
-    updatedArray.splice(index, 1);
-  }
-  return updatedArray;
-}
-
-function arrayInsert<T>(array: T[], item: T, index: number) {
-  return [...array.slice(0, index), item, ...array.slice(index)];
-}
-
-/**
- * Wraps an array around itself at a given start index
- * Example: `wrapArray(['a', 'b', 'c', 'd'], 2) === ['c', 'd', 'a', 'b']`
- */
-function wrapArray<T>(array: T[], startIndex: number) {
-  return array.map((_, index) => array[(startIndex + index) % array.length]);
-}
-
-export { arrayRemove, arrayInsert, wrapArray };
