+++ packages/react/menu/src/useMenuTypeahead.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { getSelector, wrapArray } from '@radix-ui/utils';
+import { getSelector } from '@radix-ui/utils';
 
 function useMenuTypeahead() {
   const timerRef = React.useRef(0);
@@ -82,6 +82,14 @@ type UseMenuTypeaheadItemOptions = { textValue: string; disabled?: boolean };
 
 function useMenuTypeaheadItem({ textValue, disabled }: UseMenuTypeaheadItemOptions) {
   return { [ITEM_ATTR]: disabled ? undefined : textValue };
+}
+
+/**
+ * Wraps an array around itself at a given start index
+ * Example: `wrapArray(['a', 'b', 'c', 'd'], 2) === ['c', 'd', 'a', 'b']`
+ */
+function wrapArray<T>(array: T[], startIndex: number) {
+  return array.map((_, index) => array[(startIndex + index) % array.length]);
 }
 
 export { useMenuTypeahead, useMenuTypeaheadItem };
